Adds update-cluster-membership command to op-conductor-ops (#4086)
