+++ op-conductor-ops/op-conductor-ops
@@ -1,3 +1,3 @@
 #!/bin/sh
 
-poetry run python op-conductor-ops.py ${@}
+poetry run python op-conductor-ops.py "${@}"
